clean up ; simplify code structre
